test(IT-Cotato#225): 상품 스크랩 취소 테스트 구현
- ProductServiceTest에서 스크랩 관련 테스트를 ProductScrapTest로 분리
- 스크랩 취소 성공
- 스크랩되지 않은 상품 스크랩 취소 실패
