+++ src/test/java/com/cotato/kampus/domain/product/application/ProductServiceTest.java
@@ -39,43 +39,37 @@ import com.cotato.kampus.global.util.s3.S3Uploader;
 import com.cotato.kampus.helper.TestUserHelper;
 
 @ExtendWith(MockitoExtension.class)
-class ProductServiceTest {
+public class ProductServiceTest {
 
 	@InjectMocks
-	private ProductService productService;
+	protected ProductService productService;
 
 	@Mock
-	private ProductSaver productSaver;
+	protected ProductSaver productSaver;
 
 	@Mock
-	private ProductPhotoAppender productPhotoAppender;
+	protected ProductPhotoAppender productPhotoAppender;
 
 	@Mock
-	private S3Uploader s3Uploader;
+	protected S3Uploader s3Uploader;
 
 	@Mock
-	private ApiUserResolver apiUserResolver;
+	protected ApiUserResolver apiUserResolver;
 
 	@Mock
-	private UserValidator userValidator;
+	protected UserValidator userValidator;
 
 	@Mock
-	private ImageValidator imageValidator;
+	protected ImageValidator imageValidator;
 
 	@Mock
-	private ProductCategoryFinder productCategoryFinder;
+	protected ProductCategoryFinder productCategoryFinder;
 
 	@Mock
-	private ProductCategoryMappingAdapter productCategoryMappingAdapter;
+	protected ProductCategoryMappingAdapter productCategoryMappingAdapter;
 
 	@Mock
-	private ProductFinder productFinder;
-
-	@Mock
-	private ProductScrapFinder productScrapFinder;
-
-	@Mock
-	private ProductScrapManager productScrapManager;
+	protected ProductFinder productFinder;
 
 	@Nested
 	@DisplayName("상품 생성 성공 테스트")
@@ -359,98 +353,5 @@ class ProductServiceTest {
 
 			then(productSaver).should(never()).update(any(Product.class));
 		}
-	}
-
-	@Nested
-	@DisplayName("상품 스크랩 테스트")
-	class ScrapeProductTest {
-
-		@Test
-		@DisplayName("상품 스크랩 성공")
-		void addScrap_success() {
-			// Given: 유효한 상품과 사용자가 주어졌을 때
-			Long productId = 100L;
-			Long userId = 1L;
-			UserDto user = TestUserHelper.createUserDto(userId, 1L, UserRole.VERIFIED);
-
-			Product product = Product.create(userId, "빈티지 카메라", 10000, "상태 좋아요!");
-
-			Product scrappedProduct = product.increaseChatCount();
-
-			given(apiUserResolver.getCurrentUserDto()).willReturn(user);
-			given(userValidator.validateStudentVerification(user)).willReturn(user.universityId());
-			given(productFinder.findById(productId)).willReturn(product);
-			given(productScrapFinder.isScrapped(productId, userId)).willReturn(false);
-			given(productSaver.update(any(Product.class))).willReturn(scrappedProduct);
-			willDoNothing().given(productScrapManager).append(productId, userId);
-
-			// When
-			productService.addScrap(productId);
-
-			// Then: 스크랩 추가되고 모든 의존성 호출
-			then(apiUserResolver).should().getCurrentUserDto();
-			then(userValidator).should().validateStudentVerification(user);
-			then(productFinder).should().findById(productId);
-			then(productScrapFinder).should().isScrapped(productId, userId);
-			then(productScrapManager).should().append(productId, userId);
-			then(productSaver).should().update(argThat(updatedProduct ->
-				updatedProduct.getScrapCount() == product.getScrapCount() + 1));
-		}
-
-		@Test
-		@DisplayName("삭제된 상품 스크랩 실패")
-		void addScrap_fail_deletedProduct() {
-			// Given: 삭제된 상품일 때
-			Long productId = 100L;
-			Long userId = 1L;
-			UserDto user = TestUserHelper.createUserDto(userId, 1L, UserRole.VERIFIED);
-
-			Product deletedProduct = Product.create(
-				userId, "빈티지 카메라", 10000, "상태 좋아요!")
-				.withProductStatus(ProductStatus.DELETED);
-
-			given(apiUserResolver.getCurrentUserDto()).willReturn(user);
-			given(userValidator.validateStudentVerification(user)).willReturn(user.universityId());
-			given(productFinder.findById(productId)).willReturn(deletedProduct);
-
-			// When & Then
-			assertThatThrownBy(() -> productService.addScrap(productId))
-				.isInstanceOf(AppException.class)
-				.hasMessage(ErrorCode.ALREADY_DELETED_PRODUCT.getMessage());
-
-			then(productScrapFinder).should(never()).isScrapped(anyLong(), anyLong());
-			then(productScrapManager).should(never()).append(anyLong(), anyLong());
-			then(productSaver).should(never()).update(any(Product.class));
-		}
-
-		@Test
-		@DisplayName("이미 스크랩한 상품 스크랩 실패")
-		void addScrap_fail_alreadyScrapped() {
-			// Given: 이미 스크랩한 상품일 때
-			Long productId = 100L;
-			Long userId = 1L;
-			UserDto user = TestUserHelper.createUserDto(userId, 1L, UserRole.VERIFIED);
-
-			Product product = Product.create(
-				userId,
-				"빈티지 카메라",
-				10000,
-				"상태 좋아요!"
-			);
-
-			given(apiUserResolver.getCurrentUserDto()).willReturn(user);
-			given(userValidator.validateStudentVerification(user)).willReturn(user.universityId());
-			given(productFinder.findById(productId)).willReturn(product);
-			given(productScrapFinder.isScrapped(productId, userId)).willReturn(true);
-
-			// When & Then: 예외 발생
-			assertThatThrownBy(() -> productService.addScrap(productId))
-				.isInstanceOf(AppException.class)
-				.hasMessage(ErrorCode.ALREADY_SCRAPPED_PRODUCT.getMessage());
-
-			then(productScrapManager).should(never()).append(anyLong(), anyLong());
-			then(productSaver).should(never()).update(any(Product.class));
-		}
-
 	}
 }
